backend/requirements.txt
# Async file operations
aiofiles>=23.2.1
# Core Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-dotenv>=1.0.0

# AI/ML Libraries
openai>=1.13.3
tiktoken>=0.5.0

# GNN Libraries (Optional - for weeks 9-12)
torch>=2.0.0
torch-geometric>=2.3.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0  # Updated for Python 3.12 compatibility
networkx>=3.2.0
scikit-learn>=1.3.0

# Vector Search
faiss-cpu>=1.7.4

# Utilities
httpx>=0.25.0
requests>=2.31.0
structlog>=23.2.0
pyyaml>=6.0
pydantic-settings>=2.0.0
jinja2>=3.1.2

# Azure SDK Libraries
azure-storage-blob>=12.19.0
azure-search-documents>=11.4.0
azure-cosmos>=4.5.0
azure-ai-ml>=1.12.0
azure-identity>=1.15.0

# Azure Cognitive Services - Text Analytics
azure-ai-textanalytics>=5.3.0

# Azure Monitor and Application Insights
azure-monitor-opentelemetry>=1.2.0

# Azure Core for credential management
azure-core>=1.29.0

# Graph Database Libraries

# Use versions compatible with Cosmos DB Gremlin API and aiohttp
gremlinpython>=3.6.6
aiohttp>=3.9.0

# Testing
pytest>=7.4.0

# Experiment Tracking & Optimization
optuna>=3.0.0
wandb>=0.16.0
matplotlib>=3.7.0
seaborn>=0.12.0
